requirements.txt
flask==3.0.3
python-dotenv==1.1.1
pinecone-client==7.3.0
anthropic==0.62.0
mistralai==1.9.3
langchain==0.3.0
